vz7: parasite -- Use long for memory size
When checkpointing applications with really big
memory slab (like in our vz7 test with 920G of
memory) the int type get cutted, we should use
long int instead, just like we do in other code
pieces.

Otherwise get

 | pie: 756: Daemon waits for command
 | (01.193097) Wait for ack 12 on daemon socket
 | (01.193112) Fetched ack: 12 12 0
 | (01.193164) 988065 fdinfo 0: pos:                0 flags:           100002/0
 | (01.193201) fdinfo: type: 0xb flags: 0100002/0 pos:        0 fd: 0
 | (01.193279) 988065 fdinfo 1: pos:                0 flags:           100002/0
 | (01.193307) fdinfo: type: 0xb flags: 0100002/0 pos:        0 fd: 1
 | (01.193341) 988065 fdinfo 2: pos:                0 flags:           100002/0
 | (01.193365) fdinfo: type: 0xb flags: 0100002/0 pos:        0 fd: 2
 | (01.193375) ----------------------------------------
 | (01.193405) Error (criu/parasite-syscall.c:243): BUG at criu/parasite-syscall.c:243
 | pie: 756: Error (criu/pie/parasite.c:676): Trimmed message received (1>

https://jira.sw.ru/browse/PSBM-61961

+++ criu/include/parasite-syscall.h
@@ -77,7 +77,7 @@ extern int parasite_dump_posix_timers_seized(struct proc_posix_timers_stat *proc
 		ctl->addr_args;						\
 	})
 
-extern void *parasite_args_s(struct parasite_ctl *ctl, int args_size);
+extern void *parasite_args_s(struct parasite_ctl *ctl, unsigned long args_size);
 extern int parasite_send_fd(struct parasite_ctl *ctl, int fd);
 
 /*
